pyproject.toml
[project]
name = "cold-ai"
version = "0.1.0"
description = "Phase 1 MVP for agentic cold outreach"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "jinja2>=3.1.4",
  "python-dateutil>=2.9.0.post0",
]

[project.scripts]
cold-ai = "cold_ai.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
